[Fix bugs]

+++ AniSkelF14/beziercurveevaluator.cpp
@@ -8,7 +8,7 @@ void BezierCurveEvaluator::evaluateCurve(const std::vector<Point>& control_point
 	evaluated_curve_points.clear();
 
 	if (control_points.size() < 4) {
-		for (int i = 0; i < control_points.size(); i++) {
+		for (unsigned int i = 0; i < control_points.size(); i++) {
 			evaluated_curve_points.push_back(control_points[i]);
 		}
 		
@@ -33,7 +33,7 @@ void BezierCurveEvaluator::evaluateCurve(const std::vector<Point>& control_point
 		evaluated_curve_points.push_back(Point(x, y));
 
 	} else {
-		const Mat4d M(-1, 3, -3, 1,
+		const Mat4f M(-1, 3, -3, 1,
 					   3, -6, 3, 0,
 				      -3, 3, 0, 0,
 					   1, 0, 0, 0);
@@ -44,7 +44,7 @@ void BezierCurveEvaluator::evaluateCurve(const std::vector<Point>& control_point
 		}
 
 		if (wrap_control_points) {
-			for (int i = 0; i < 3; i++) {
+			for (unsigned int i = 0; i < 3; i++) {
 				controlPoints.push_back(Point(control_points[0].x + animation_length, control_points[0].y));
 				controlPoints.insert(controlPoints.begin(), Point(control_points[control_points.size() - 1].x - animation_length, control_points[control_points.size() - 1].y));
 			}
@@ -56,16 +56,16 @@ void BezierCurveEvaluator::evaluateCurve(const std::vector<Point>& control_point
 		}*/
 
 		int numberOfControlPoint = controlPoints.size();
-		for (int i = 0; i < numberOfControlPoint - 3; i += 3) {
+		for (unsigned int i = 0; i < numberOfControlPoint - 3; i += 3) {
 			Point points[4];
-			for (int j = 0; j < 4; j++) {
+			for (unsigned int j = 0; j < 4; j++) {
 				points[j] = controlPoints[i + j];
 			}
-			Vec4d Px(points[0].x, points[1].x, points[2].x, points[3].x);
-			Vec4d Py(points[0].y, points[1].y, points[2].y, points[3].y);
-			for (int j = 0; j < 100; j++) {
+			Vec4f Px(points[0].x, points[1].x, points[2].x, points[3].x);
+			Vec4f Py(points[0].y, points[1].y, points[2].y, points[3].y);
+			for (unsigned int j = 0; j < 100; j++) {
 				float t = j / 100.0f;
-				const Vec4d T(t*t*t, t*t, t, 1);
+				const Vec4f T(t*t*t, t*t, t, 1);
 				Point Q;
 				Q.x = T * M * Px;
 				Q.y = T * M * Py;
